Rename packages to reflect new modid
