+++ gradle.properties
@@ -3,15 +3,15 @@
 gtnh.settings.blowdryerTag = 0.2.2
 
 # Human-readable mod name, available for mcmod.info population.
-modName = Pressurized Essentia Pipes
+modName = Essentia Pipes
 
 # Case-sensitive identifier string, available for mcmod.info population and used for automatic mixin JSON generation.
 # Conventionally lowercase.
-modId = pressurizedessentia
+modId = essentiapipes
 
 # Root package of the mod, used to find various classes in other properties,
 # mcmod.info substitution, enabling assertions in run tasks, etc.
-modGroup = dev.rndmorris.pressurizedessentia
+modGroup = dev.rndmorris.essentiapipes
 
 # Whether to use modGroup as the maven publishing group.
 # Due to a history of using JitPack, the default is com.github.GTNewHorizons for all mods.
@@ -50,7 +50,7 @@ enableGenericInjection = true
 # Generate a class with a String field for the mod version named as defined below.
 # If generateGradleTokenClass is empty or not missing, no such class will be generated.
 # If gradleTokenVersion is empty or missing, the field will not be present in the class.
-generateGradleTokenClass = dev.rndmorris.pressurizedessentia.Tags
+generateGradleTokenClass = dev.rndmorris.essentiapipes.Tags
 
 # Name of the token containing the project's current version to generate/replace.
 gradleTokenVersion = VERSION
