+++ src/main/java/dev/rndmorris/essentiapipes/blocks/BlockPipeSegment.java
@@ -1,4 +1,4 @@
-package dev.rndmorris.pressurizedessentia.blocks;
+package dev.rndmorris.essentiapipes.blocks;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -23,16 +23,16 @@ import net.minecraftforge.common.util.ForgeDirection;
 import cpw.mods.fml.common.registry.GameRegistry;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
-import dev.rndmorris.pressurizedessentia.Config;
-import dev.rndmorris.pressurizedessentia.EssentiaPipes;
-import dev.rndmorris.pressurizedessentia.api.IIOPipeSegment;
-import dev.rndmorris.pressurizedessentia.api.IPipeSegment;
-import dev.rndmorris.pressurizedessentia.api.PipeColor;
-import dev.rndmorris.pressurizedessentia.api.PipeHelper;
-import dev.rndmorris.pressurizedessentia.api.WorldCoordinate;
-import dev.rndmorris.pressurizedessentia.client.BlockPipeSegmentRenderer;
-import dev.rndmorris.pressurizedessentia.items.ItemBlockPipeSegment;
-import dev.rndmorris.pressurizedessentia.tile.TileEntityIOPipeSegment;
+import dev.rndmorris.essentiapipes.Config;
+import dev.rndmorris.essentiapipes.EssentiaPipes;
+import dev.rndmorris.essentiapipes.api.IIOPipeSegment;
+import dev.rndmorris.essentiapipes.api.IPipeSegment;
+import dev.rndmorris.essentiapipes.api.PipeColor;
+import dev.rndmorris.essentiapipes.api.PipeHelper;
+import dev.rndmorris.essentiapipes.api.WorldCoordinate;
+import dev.rndmorris.essentiapipes.client.BlockPipeSegmentRenderer;
+import dev.rndmorris.essentiapipes.items.ItemBlockPipeSegment;
+import dev.rndmorris.essentiapipes.tile.TileEntityIOPipeSegment;
 import thaumcraft.api.aspects.IEssentiaTransport;
 import thaumcraft.api.wands.IWandable;
 import thaumcraft.codechicken.lib.raytracer.IndexedCuboid6;
