+++ src/main/resources/assets/essentiapipes/lang/en_US.lang
@@ -1,0 +1,5 @@
+tile.essentiapipes:pipe_segment.name=Essentia Pipe
+tile.essentiapipes:pipe_segment_thaumium.name=Thaumium-Reinforced Essentia Pipe
+tile.essentiapipes:pipe_segment_voidmetal.name=Voidmetal-Reinforced Essentia Pipe
+
+itemGroup.essentiapipes=Essentia Pipes
