+++ src/main/resources/assets/essentiapipes/textures/blocks/pipe_segment_0.png
