+++ src/main/resources/assets/essentiapipes/textures/blocks/pipe_segment_1.png
