+++ src/main/resources/assets/essentiapipes/textures/blocks/pipe_segment_thaumium_3.png
