+++ src/main/resources/assets/essentiapipes/textures/blocks/pipe_segment_voidmetal_0.png
