+++ src/main/resources/assets/essentiapipes/textures/blocks/pipe_segment_voidmetal_5.png
